chore(web): reorganize project services into lib/projects/ module
Move projectsService.ts → lib/projects/svc.ts, useProjects.ts → lib/projects/hooks.ts,
and extract all types/interfaces/enums into lib/projects/types.ts. Update all import
sites accordingly.
