+++ web/src/providers/ProjectsContext.tsx
@@ -14,12 +14,14 @@ import {
 } from "react";
 import useSWR from "swr";
 import { errorHandlingFetcher, skipRetryOnAuthError } from "@/lib/fetcher";
-import type {
-  CategorizedFiles,
-  Project,
-  ProjectFile,
-  UserFileDeleteResult,
-} from "@/app/app/projects/projectsService";
+import {
+  UserFileStatus,
+  type CategorizedFiles,
+  type Project,
+  type ProjectDetails,
+  type ProjectFile,
+  type UserFileDeleteResult,
+} from "@/lib/projects/types";
 import {
   fetchProjects as svcFetchProjects,
   createProject as svcCreateProject,
@@ -30,24 +32,22 @@ import {
   getProjectInstructions as svcGetProjectInstructions,
   upsertProjectInstructions as svcUpsertProjectInstructions,
   getProjectDetails as svcGetProjectDetails,
-  ProjectDetails,
   renameProject as svcRenameProject,
   deleteProject as svcDeleteProject,
   deleteUserFile as svcDeleteUserFile,
   getUserFileStatuses as svcGetUserFileStatuses,
   unlinkFileFromProject as svcUnlinkFileFromProject,
   linkFileToProject as svcLinkFileToProject,
-  UserFileStatus,
-} from "@/app/app/projects/projectsService";
+} from "@/lib/projects/svc";
 import { useSearchParams } from "next/navigation";
 import { SEARCH_PARAM_NAMES } from "@/app/app/services/searchParams";
 import { useAppRouter } from "@/hooks/appNavigation";
 import { ChatFileType } from "@/app/app/interfaces";
 import { toast } from "@/hooks/useToast";
-import { useProjects } from "@/lib/hooks/useProjects";
+import { useProjects } from "@/lib/projects/hooks";
 import { useSettings } from "@/lib/settings/hooks";
 
-export type { Project, ProjectFile } from "@/app/app/projects/projectsService";
+export type { Project, ProjectFile } from "@/lib/projects/types";
 
 // Helper to generate unique temp IDs
 const generateTempId = () => {
